+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Job result files listing and download endpoints in Admin UI
+*
 
 ### Changed
 
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+
+## [0.10.0] - 2026-03-10
+
+### Added
+
+* Job result files listing and download endpoints in Admin UI
 
 ## [0.9.0] - 2026-03-10